+++ Assets/02. Scripts/Runtime/Enemies/ViewControllers/Instances/Boss1/Boss1.cs
@@ -49,7 +49,7 @@ namespace Runtime.Enemies
             return level * baseWeight;
         }
 
-        public override int OnGetRealSpawnCost(int level, int rarity, int baseCost) {
+        public override float OnGetRealSpawnCost(int level, int rarity, float baseCost) {
             return level * baseCost;
         }
