+++ Assets/02. Scripts/Runtime/Levels/Models/Base/LevelEntity.cs
@@ -13,13 +13,13 @@ using Runtime.Utilities.ConfigSheet;
 
 namespace _02._Scripts.Runtime.Levels.Models {
 	public interface ILevelEntity : IEntity, IHaveCustomProperties, IHaveTags {
-		public List<LevelSpawnCard> GetAllCardsUnderCost(int cost);
+		public List<LevelSpawnCard> GetAllCardsUnderCost(float cost);
 		
-		public List<LevelSpawnCard> GetAllCardsUnderCost(int cost, Predicate<LevelSpawnCard> furtherPredicate);
+		public List<LevelSpawnCard> GetAllCardsUnderCost(float cost, Predicate<LevelSpawnCard> furtherPredicate);
 		
-		public List<LevelSpawnCard> GetAllNormalEnemiesUnderCost(int cost);
+		public List<LevelSpawnCard> GetAllNormalEnemiesUnderCost(float cost);
 		
-		public List<LevelSpawnCard> GetAllNormalEnemiesUnderCost(int cost, Predicate<LevelSpawnCard> furtherPredicate);
+		public List<LevelSpawnCard> GetAllNormalEnemiesUnderCost(float cost, Predicate<LevelSpawnCard> furtherPredicate);
 
 		public List<LevelSpawnCard> GetAllBosses();
 
@@ -68,7 +68,7 @@ namespace _02._Scripts.Runtime.Levels.Models {
 			return card.MinRarity;
 		}
 		
-		public List<LevelSpawnCard> GetAllCardsUnderCost(int cost) {
+		public List<LevelSpawnCard> GetAllCardsUnderCost(float cost) {
 			List<LevelSpawnCard> cards = new List<LevelSpawnCard>();
 			int level = GetCurrentLevelCount();
 			foreach (var card in spawnCardsProperty.RealValues) {
@@ -80,7 +80,7 @@ namespace _02._Scripts.Runtime.Levels.Models {
 			return cards;
 		}
 
-		public List<LevelSpawnCard> GetAllCardsUnderCost(int cost, Predicate<LevelSpawnCard> furtherPredicate) {
+		public List<LevelSpawnCard> GetAllCardsUnderCost(float cost, Predicate<LevelSpawnCard> furtherPredicate) {
 			List<LevelSpawnCard> cards = new List<LevelSpawnCard>();
 			int level = GetCurrentLevelCount();
 			foreach (var card in spawnCardsProperty.RealValues) {
@@ -91,12 +91,12 @@ namespace _02._Scripts.Runtime.Levels.Models {
 			return cards;
 		}
 
-		public List<LevelSpawnCard> GetAllNormalEnemiesUnderCost(int cost) {
+		public List<LevelSpawnCard> GetAllNormalEnemiesUnderCost(float cost) {
 			return GetCards((card =>
 				card.GetRealSpawnCost(GetCurrentLevelCount(), GetMinRarity(card)) <= cost && card.IsNormalEnemy));
 		}
 
-		public List<LevelSpawnCard> GetAllNormalEnemiesUnderCost(int cost, Predicate<LevelSpawnCard> furtherPredicate) {
+		public List<LevelSpawnCard> GetAllNormalEnemiesUnderCost(float cost, Predicate<LevelSpawnCard> furtherPredicate) {
 			return GetCards((card =>
 				card.GetRealSpawnCost(GetCurrentLevelCount(), GetMinRarity(card)) <= cost && card.IsNormalEnemy &&
 				furtherPredicate(card)));
